+++ Assets/1_Scripts/Networking/GameManager.cs
@@ -17,14 +17,7 @@ namespace Networking
         [Tooltip("The prefab to use for representing the player")]
         public GameObject playerPrefab;
         public GameObject[] spawnPoints;
-        public GameObject personalMenuCanvas;
-        public float timer;
         
-        protected Vector3 previousPosition;
-        protected Vector3 newPosition;
-        protected Vector3 playerFirstPos;
-        
-        private Transform headsetTransform;
 
         
         private void Start()
@@ -37,14 +30,8 @@ namespace Networking
             {
                 Debug.LogFormat("We are Instantiating LocalPlayer from {0}", SceneManager.GetActiveScene().name);
                 // we're in a room. spawn a character for the local player. it gets synced by using PhotonNetwork.Instantiate
-                playerFirstPos = spawnPoints[Random.Range(0, spawnPoints.Length - 1)].transform.position;
-                PhotonNetwork.Instantiate(this.playerPrefab.name, playerFirstPos, Quaternion.identity, 0);
+                PhotonNetwork.Instantiate(this.playerPrefab.name, spawnPoints[Random.Range(0, spawnPoints.Length - 1)].transform.position, Quaternion.identity, 0);
             }
-
-            //on initialise des données qui seront traitées dans l'update()
-            previousPosition = playerFirstPos;
-            newPosition = new Vector3(0,0,0);
-            timer = 0.0f;
         }
         
         
@@ -97,77 +84,6 @@ namespace Networking
 
         #endregion
         
-        
-        //fonction qui définit s'il y a un l'utilisateur se déplace à partir de ses positions
-        protected bool userMoves(Vector3 previousPosition, Vector3 newposition)
-        {
-            return (!(Vector3.Magnitude(previousPosition - newposition) <= 0.0003f));
-        }
-        
-        //Fonction qui change la valeur du statut de l'utilisateur, selon depuis combien de temps il n'a pas bougé (passage de connecté à absent)
-        //où selon s'il est dans un menu
-        protected void updateStatus(float timer, Vector3 previousPosition, Vector3 newPosition, out float updatedTimer)
-        {
-            //le string sur lequel on va calculer le statut
-            //on attribue la valeur en sortie du timer au cas où elle ne serait pas changée
-            //Si l'utilisateur bouge, il est soit connecté soit dans un menu
-            //Sinon, s'il ne bouge pas pendant plus de 120s, il est considéré absent
-            //On attribue finalementà la playerpref userStatus sa nouvelle valeur
-
-            string status = "connecté";
-            updatedTimer = timer;
-            
-            
-            if (userMoves(previousPosition, newPosition))
-            {
-                updatedTimer = 0;
-                status = "connecté";
-            }
-            if (personalMenuCanvas.activeInHierarchy)
-            {
-                status = "dans le menu";
-            }
-            if (timer >= 120.0f)
-            {
-                status = "absent";
-            }
-            
-            
-            UnityEngine.PlayerPrefs.SetString("userStatus", status);
-        }
-        
-        
-
-        private void Update()
-        {
-            //dans l'update, on va évaluer le statut de l'utilisateur, 
-            //pour celà, on va réévaluer l'ancienne position previousPosition et la nouvelle newPosition
-            //On pourra ensuite les traiter à l'aide de la fonction updateStatus, qui nous permettra de
-            //savoir avoec le timer si l'utilisateur est absent, connecté, où dans le menu
-            
-            previousPosition = newPosition; 
-            if (headsetTransform != null)
-            {
-                newPosition = headsetTransform.position;
-            }
-            else
-            {
-                try
-                {
-                    headsetTransform = VRTK_DeviceFinder.DeviceTransform(VRTK_DeviceFinder.Devices.Headset).gameObject
-                        .transform;
-                    newPosition = headsetTransform.position;
-                }
-                catch (NullReferenceException)
-                {
-                    newPosition = Vector3.zero;
-                }
-            }
-
-            //Avec les lignes suivantes, on fait évoluer le timer qui compte la durée de non-mouvement de l'utilisateur, et on fait évoluer le statut
-            timer += Time.deltaTime;
-            updateStatus(timer, previousPosition, newPosition, out timer);
-        }
     }
  
 }
