+++ Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -37,5 +37,6 @@ MonoBehaviour:
   - SwipeLeft
   - SwipeRight
   - updateUserStatus
+  - updateStatusSphereMaterial
   DisableAutoOpenWizard: 1
   ShowSettings: 1
