+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - VRTKManager
   - IncludeTeleport
   - LabelToggleUser
+  - PersonnalMenu
   layers:
   - Default
   - TransparentFX
